requirements.txt
filterpy>=1.4.5
ftfy>=6.1.1
gdown==5.1.0
GitPython>=3.1.0
lapx>=0.5.4
loguru>=0.7.0
matplotlib>=3.3.0
# numpy>=1.23.0
numpy==1.24.4
opencv-python>=4.6.0
pandas>=1.1.4
pillow>=7.1.2
pre-commit>=3.3.3
psutil
py-cpuinfo
PyYAML>=5.3.1
regex>=2023.6.3
requests>=2.23.0
scikit-learn>=1.3.0
scipy>=1.4.1
seaborn>=0.11.0
tensorboard>=2.13.0
# torch>=1.7.0
torch>=1.8.0
# torchvision>=0.8.1
torchvision>=0.9.0
tqdm>=4.64.0
ultralytics-thop>=2.0.0
yacs>=0.1.8
ultralytics==8.4.5
boxmot==15.0.10


# Export ----------------------------------------------------------------------

# onnx>=1.12.0          # ONNX export
# onnxsim>=0.4.1        # ONNX simplifier
# nvidia-pyindex        # TensorRT export
# nvidia-tensorrt       # TensorRT export
# openvino-dev>=2022.3  # OpenVINO export
# onnx2tf>=1.10.0       # TFLite export

# Hyperparam search -----------------------------------------------------------

# optuna                # genetic algo
# plotly                # hyper param importance and pareto front plots
# kaleido
# joblib
